requirements.txt
colorama==0.4.1
cycler==0.10.0
Django==2.1.5
django-extra-fields==1.0.0
django-filter==2.1.0
django-filters==0.2.1
django-unixtimestampfield==0.3.9
djangorestframework==3.9.1
kiwisolver==1.0.1
matplotlib==3.0.3
numpy==1.14.5
opencv-python==4.0.0.21
pandas==0.24.2
Pillow==5.1.0
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2018.9
pyzbar==0.1.8
qrcode==6.1
scipy==1.2.1
Shapely==1.6.4.post1
six==1.12.0
xlrd==1.2.0
XlsxWriter==1.1.8
